Move IntegrationTestbaseVideo into baseclass.py and prepare for audio-comparison tests
