+++ python-api/tests/integrationtests2/test_video.py
@@ -4,34 +4,9 @@ Integration Tests for the Video-Pipelines
 
 from __future__ import absolute_import, print_function, unicode_literals
 
-from .baseclass import IntegrationTestbase, IntegrationTestbaseCompare
-from gstswitch.testsource import VideoSrc
+from .baseclass import IntegrationTestbaseVideo
 from gstswitch.controller import Controller
 import pytest
-
-
-class IntegrationTestbaseVideo(IntegrationTestbaseCompare):
-    """ Testbase used to test video-pipeline results
-    """
-
-    def setup_test(self):
-        """Setup Server, Controller and two Video-Test-Sources"""
-        self.setup_server()
-        self.setup_controller()
-
-        self.log.info("starting up 1st VideoSrc")
-        self.new_test_video(pattern=VideoSrc.PATTERN_RED)
-
-        self.log.info("starting up 2nd VideoSrc")
-        self.new_test_video(pattern=VideoSrc.PATTERN_GREEN)
-
-        self.log.info("starting up 3rd VideoSrc")
-        self.new_test_video(pattern=VideoSrc.PATTERN_BLUE)
-
-        self.serv.wait_for_output('tcpserversink name=sink', count=3)
-
-        self.log.info("waiting for end of initial transition")
-        self.serv.wait_for_output('ending transition')
 
 
 class TestCompositeModes(IntegrationTestbaseVideo):
@@ -42,50 +17,50 @@ class TestCompositeModes(IntegrationTestbaseVideo):
         """Test set_composite_mode COMPOSITE_NONE"""
         self.setup_test()
 
-        self.expect_frame('COMPOSITE_DUAL_EQUAL.png')
+        self.expect_video_frame('COMPOSITE_DUAL_EQUAL.png')
 
         self.log.info("setting composite-mode=COMPOSITE_NONE")
         assert self.controller.set_composite_mode(
             Controller.COMPOSITE_NONE)
 
-        self.expect_frame('COMPOSITE_NONE.png')
+        self.expect_video_frame('COMPOSITE_NONE.png')
 
     def test_set_composite_mode_pip(self):
         """Test set_composite_mode COMPOSITE_PIP"""
         self.setup_test()
 
-        self.expect_frame('COMPOSITE_DUAL_EQUAL.png')
+        self.expect_video_frame('COMPOSITE_DUAL_EQUAL.png')
 
         self.log.info("setting composite-mode=COMPOSITE_PIP")
         assert self.controller.set_composite_mode(
             Controller.COMPOSITE_PIP)
 
-        self.expect_frame('COMPOSITE_PIP.png')
+        self.expect_video_frame('COMPOSITE_PIP.png')
 
     def test_set_composite_mode_preview(self):
         """Test set_composite_mode COMPOSITE_DUAL_PREVIEW"""
         self.setup_test()
 
-        self.expect_frame('COMPOSITE_DUAL_EQUAL.png')
+        self.expect_video_frame('COMPOSITE_DUAL_EQUAL.png')
 
         self.log.info("setting composite-mode=COMPOSITE_DUAL_PREVIEW")
         assert self.controller.set_composite_mode(
             Controller.COMPOSITE_DUAL_PREVIEW)
 
-        self.expect_frame('COMPOSITE_DUAL_PREVIEW.png')
+        self.expect_video_frame('COMPOSITE_DUAL_PREVIEW.png')
 
     def test_set_composite_mode_equal(self):
         """Test set_composite_mode COMPOSITE_DUAL_EQUAL"""
         self.setup_test()
 
-        self.expect_frame('COMPOSITE_DUAL_EQUAL.png')
+        self.expect_video_frame('COMPOSITE_DUAL_EQUAL.png')
 
         # mode is unchanged -> returns false
         self.log.info("setting composite-mode=COMPOSITE_DUAL_EQUAL")
         assert not self.controller.set_composite_mode(
             Controller.COMPOSITE_DUAL_EQUAL)
 
-        self.expect_frame('COMPOSITE_DUAL_EQUAL.png')
+        self.expect_video_frame('COMPOSITE_DUAL_EQUAL.png')
 
 
 class TestSwitch(IntegrationTestbaseVideo):
@@ -100,49 +75,49 @@ class TestSwitch(IntegrationTestbaseVideo):
         self.setup_test()
 
         # default: A=RED, B=GREEN
-        self.expect_frame('SWITCH_RED_GREEN.png')
+        self.expect_video_frame('SWITCH_RED_GREEN.png')
 
         # switch: A=GREEN, [B=RED]
         self.log.info("switching video channel A to GREEN")
         assert self.controller.switch(
             Controller.VIDEO_CHANNEL_A, self.PORT_GREEN)
 
-        self.expect_frame('SWITCH_GREEN_RED.png')
+        self.expect_video_frame('SWITCH_GREEN_RED.png')
 
     def test_switch_video_source_b(self):
         """Test set_composite_mode COMPOSITE_NONE"""
         self.setup_test()
 
         # default: A=RED, B=GREEN
-        self.expect_frame('SWITCH_RED_GREEN.png')
+        self.expect_video_frame('SWITCH_RED_GREEN.png')
 
         # switch: [A=GREEN], B=RED
         self.log.info("switching video channel B to RED")
         assert self.controller.switch(
             Controller.VIDEO_CHANNEL_B, self.PORT_RED)
 
-        self.expect_frame('SWITCH_GREEN_RED.png')
+        self.expect_video_frame('SWITCH_GREEN_RED.png')
 
     def test_switch_video_source_ab(self):
         """Test set_composite_mode COMPOSITE_NONE"""
         self.setup_test()
 
         # default: A=RED, B=GREEN
-        self.expect_frame('SWITCH_RED_GREEN.png')
+        self.expect_video_frame('SWITCH_RED_GREEN.png')
 
         # switch: A=BLUE, B=GREEN
         self.log.info("switching video channel A to BLUE")
         assert self.controller.switch(
             Controller.VIDEO_CHANNEL_A, self.PORT_BLUE)
 
-        self.expect_frame('SWITCH_BLUE_GREEN.png')
+        self.expect_video_frame('SWITCH_BLUE_GREEN.png')
 
         # switch: B=BLUE, [A=GREEN]
         self.log.info("switching video channel B to BLUE")
         assert self.controller.switch(
             Controller.VIDEO_CHANNEL_B, self.PORT_BLUE)
 
-        self.expect_frame('SWITCH_GREEN_BLUE.png')
+        self.expect_video_frame('SWITCH_GREEN_BLUE.png')
 
 
 class TestAdjustPIP(IntegrationTestbaseVideo):
@@ -158,17 +133,17 @@ class TestAdjustPIP(IntegrationTestbaseVideo):
             Controller.COMPOSITE_PIP)
 
         # default: A=RED, B=GREEN
-        self.expect_frame('COMPOSITE_PIP.png')
+        self.expect_video_frame('COMPOSITE_PIP.png')
 
         # adjusted pip: A=RED, B=GREEN
         self.log.info("calling adjust_pip")
         assert self.controller.adjust_pip(
             74, 92, 90, 60)
 
-        self.expect_frame('COMPOSITE_PIP_PLACEMENT.png')
+        self.expect_video_frame('COMPOSITE_PIP_PLACEMENT.png')
 
 
-class TestCompositeOutput(IntegrationTestbase):
+class TestCompositeOutput(IntegrationTestbaseVideo):
     """ Test the actual Output of the Composition-Port
     """
     def test_caps(self):
@@ -178,7 +153,7 @@ class TestCompositeOutput(IntegrationTestbase):
         pass
 
 
-class TestEncodedOutput(IntegrationTestbase):
+class TestEncodedOutput(IntegrationTestbaseVideo):
     """ Test the actual Output of the Endoded-Port
     """
     def test_caps(self):
